Identity event reconstruction

+++ pretrain_on_jigsaw_events.py
@@ -30,11 +30,11 @@ def train(num_points, batch_size, num_classes, num_epochs, file_stem):
     train_ds = np.load('{}{}'.format(file_stem, 'train.npy'))
     val_ds = np.load('{}{}'.format(file_stem, 'val.npy'))
     train_features = train_ds[:, :, :3]
-    train_labels = train_ds[:, :, 3]
+    train_labels = train_ds[:, :, 3].astype(int)
     train_ds = tf.data.Dataset.from_tensor_slices((train_features, train_labels)).map(fix_shape)
 
     val_features = val_ds[:, :, :3]
-    val_labels = val_ds[:, :, 3]
+    val_labels = val_ds[:, :, 3].astype(int)
     val_ds = tf.data.Dataset.from_tensor_slices((val_features, val_labels)).map(fix_shape)
     
     train_ds = train_ds.shuffle(len(train_ds)).batch(batch_size)
